decouple DeviceAdapter from Handle. Add Device Spec

+++ include/cuda/Common.h
@@ -36,7 +36,7 @@ class Kokkos: public Device {};
 
 using IDX = unsigned long long int;
 using LENGTH = unsigned long long int;
-using SIZE = uint64_t;//unsigned int;
+using SIZE = uint32_t;//unsigned int;
 // using SIZE = int;
 using DIM = uint32_t;
 using QUANTIZED_INT = int;
